Create android-logo-generator.sh
Simple script that takes a single image(logo) and creates sized copies of the relevant dpi placed into the given folders. Forsome reason this little gimmick is not easy to come by i recall seeing a npm app but cant remeber the name of it :( So here is a bash script but i cant get the rounded logo command right, its close but need to work out how to invert the alpha of the border layer.
Mogrify to make life easier.